Add policy to Pelanggan & constraint to migration(pelanggan, requestmitra, & kerjasama)

+++ app/Http/Controllers/Customer/PelangganController.php
@@ -4,15 +4,18 @@ namespace App\Http\Controllers\Customer;
 
 use App\Http\Controllers\Controller;
 use App\Models\Pelanggan;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class PelangganController extends Controller
 {
+    use AuthorizesRequests;
     public function index()
     {
         // Filters the pelanggans table to only include records where user_id matches the logged-in user's ID(There a user_id field inside table pelanggan).
-        $pelanggans = Pelanggan::where('user_id', Auth::id())->latest()->get();
+        $pelanggans = Pelanggan::where('user_id', Auth::id())
+            ->latest()->get();
 
         return view('/customer.pelanggan.index', [
             'pelanggans' => $pelanggans
@@ -58,6 +61,8 @@ class PelangganController extends Controller
 
     public function edit(Pelanggan $pelanggan)
     {
+        $this->authorize('update', $pelanggan);
+
         return view('/customer.pelanggan.edit', [
             'pelanggan' => $pelanggan
         ]);
@@ -65,12 +70,14 @@ class PelangganController extends Controller
 
     public function update(Request $request, Pelanggan $pelanggan)
     {
+        $this->authorize('update', $pelanggan);
+
         $validatedData = $request->validate([
-            'nama_pelanggan' => ['required'],
-            'jk_pelanggan' => ['required', 'in:Laki-laki,Perempuan'],
-            'noTelp_pelanggan' => ['required'],
-            'email_pelanggan' => ['required', 'email'],
-            'alamat_pelanggan' => ['required'],
+            'nama_pelanggan'    => ['required'],
+            'jk_pelanggan'      => ['required', 'in:Laki-laki,Perempuan'],
+            'noTelp_pelanggan'  => ['required'],
+            'email_pelanggan'   => ['required', 'email'],
+            'alamat_pelanggan'  => ['required'],
         ]);
 
         $authData = array_merge($validatedData, ['user_id' => Auth::id()]);
@@ -82,6 +89,8 @@ class PelangganController extends Controller
 
     public function destroy(Pelanggan $pelanggan)
     {
+        $this->authorize('delete', $pelanggan);
+
         $pelanggan->delete();
 
         return redirect('/customer/pelanggan');
